// src/api/enums/ApiErrorCodes.ts
enum ApiErrorCodes {
  UNKOWN_ERROR = "UNKOWN_ERROR",
  UNKOWN_ROUTE = "UNKOWN_ROUTE",
  UNKOWN_TEAMSPEAK_ERROR = "UNKOWN_TEAMSPEAK_ERROR",
  CHANNEL_DOES_NOT_EXIST = "CHANNEL_DOES_NOT_EXIST",
  CLIENT_DOES_NOT_EXIST = "CLIENT_DOES_NOT_EXIST",
  SERVERGROUP_DOES_NOT_EXIST = "SERVERGROUP_DOES_NOT_EXIST",
  SERVERGROUP_DUPLICATE_ENTRY = "SERVERGROUP_DUPLICATE_ENTRY",
  SERVERGROUP_EMPTY_RESULT = "SERVERGROUP_EMPTY_RESULT",
  WRONG_TYPE = "WRONG_TYPE",
  PARTIAL_ERROR = "PARTIAL_ERROR",
}

export default ApiErrorCodes;